Change endpoint location
